+++ src/devices/block/drivers/nvme/queue.h
@@ -19,7 +19,7 @@ namespace nvme {
 // The queue will always fit in one page.
 class Queue {
  public:
-  static zx::result<Queue> Create(zx::unowned_bti bti, uint16_t queue_id, size_t max_entries,
+  static zx::result<Queue> Create(zx::unowned_bti bti, uint16_t queue_id, uint32_t max_entries,
                                   size_t entry_size) {
     Queue ret(entry_size, queue_id);
     auto status = ret.Init(std::move(bti), max_entries);
@@ -32,7 +32,7 @@ class Queue {
   // Get the physical address of this queue, suitable for passing to the controller.
   zx_paddr_t GetDeviceAddress() const { return io_.phys_list()[0]; }
   // Return the number of entries in the queue.
-  size_t entry_count() const { return entry_count_; }
+  uint32_t entry_count() const { return entry_count_; }
 
   // Get the next item in the queue, and move the queue pointer forward.
   void* Next() {
@@ -58,11 +58,11 @@ class Queue {
  private:
   explicit Queue(size_t entry_size, uint16_t queue_id)
       : entry_size_(entry_size), queue_id_(queue_id) {}
-  zx::result<> Init(zx::unowned_bti bti, size_t max_entries);
+  zx::result<> Init(zx::unowned_bti bti, uint32_t max_entries);
 
   ddk::IoBuffer io_;
   size_t entry_size_;
-  size_t entry_count_;
+  uint32_t entry_count_;
   uint16_t queue_id_;
 
   size_t next_index_ = 0;
